backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-dotenv
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision